+++ extension/features/common/nav-separator.js
@@ -1,3 +1,0 @@
-export default () => {
-  return $('<div class="nav-menu-separator">&nbsp;</div>');
-};