+++ extension/features/hide-region.js
@@ -1,3 +1,6 @@
 export default () => {
-  $('#nav-regionMenu').hide();
+  const el = $('header#awsc-nav-header div[data-testid="awsc-nav-more-menu-list"] ul#awsc-navigation__more-menu--list');
+  const regions = $(el).find('button[data-testid="more-menu__awsc-nav-regions-menu-button"]');
+
+  $(regions).hide();
 };