+++ extension/features/hide-support.js
@@ -1,3 +1,6 @@
 export default () => {
-  $('#nav-helpMenu').hide();
+  const el = $('header#awsc-nav-header div[data-testid="awsc-nav-more-menu-list"] ul#awsc-navigation__more-menu--list');
+  const support = $(el).find('button[data-testid="more-menu__awsc-nav-support-menu-button"]');
+
+  $(support).hide();
 };