+++ extension/features/pin-button.js
@@ -1,7 +1,12 @@
 export default () => {
-  const el = $('#nav-shortcutMenu');
-  const parent = $(el).parent();
-  const sibling = $(parent).find('#nav-menu-right');
+  const header = $('header#awsc-nav-header');
+  const favorites = $(header).find('div[data-testid="awsc-nav-service-menu"] div[data-testid="favorites-container"]');
+  const parent = $(header).parent();
+  const recent = $(header).find('div[data-testid="awsc-nav-service-menu"] div[data-testid="awsc-nav-recently-visited-list"]');
+  const sibling = $(recent).siblings();
 
-  $(el).prependTo($(sibling));
+  recent.css({ 'margin-top': 0 });
+  $(sibling).hide();
+  $(favorites).addClass('ra-pb');
+  $(favorites).appendTo(parent);
 };