+++ extension/main.js
@@ -1,24 +1,45 @@
 import {defaultFeatures} from './features';
-import {StoreSync} from './library/store-sync';
+import {Store, StoreSync, job} from './lib/core';
+import config from './config';
 
-const storeSync = new StoreSync(window, 'refined:aws');
+const storeSync = new StoreSync(window, config.namespace.sync);
+const store = new Store(window, config.namespace.local);
 
 // Save default features option value under chrome storage sync.
 async function main() {
+  job({
+    name: 'invalidate',
+    ctx: window,
+    frequency: config.cache.frequency,
+    fn: async () => {
+      try {
+        await Promise.all([
+          store.remove(store._prefix + config.cache.servicesKey),
+          store.remove(store._prefix + config.cache.regionsKey),
+        ]);
+        console.log(config.logging.ok, 'Invalidating cache');
+      } catch (error) {
+        console.log(config.logging.error, 'Error invalidating cache');
+        console.log(error);
+      }
+    },
+  });
+
   try {
     const options = await storeSync.get(null);
     await storeSync.set({...options, ...defaultFeatures});
-    console.log('✅', 'Saved default');
+    console.log(config.logging.ok, 'Saved default');
   } catch (error) {
-    console.log('❌', 'Error saving default');
-    // Fix: Handle error, retry sync or write to log file but put under logging option.
+    console.log(config.logging.error, 'Error saving default');
     console.log(error);
   }
 }
 
 async function installerHandler({reason}) {
-  if (reason === 'install' || reason === 'update') {
-    window.chrome.runtime.openOptionsPage();
+  if (!config.devMode) {
+    if (reason === 'install' || reason === 'update') {
+      window.chrome.runtime.openOptionsPage();
+    }
   }
 }