+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+# New Version
+* Remove babel-eslint@10.1.0
+* Add chrome.alarms to clear cache periodically
+* Remove support for Move "Language Selector" button
+* Remove support for Move "Feedback" button
+* Remove support for Show "Documentation" button
+* Remove Support for Hide "Resource Groups" selection
+* Remove support for Switch Role `s, s, r` keyboard shortcut
+* Remove support for Quick Open `q, q` keyboard shortcut
+* UI changes fix
+  * Quick Favorites Navigation
+  * Region shortcuts
+  * Roles shortcuts
+  * Services shortcuts
+  * Hide Region
+  * Hide Support
+  * Show Sign Out
+* Move region mappings, namespace and cacheFrequency to config
+* Add WEB_EXT_ID to Travis file
+* Add yarn audit to test script
+* Upgrade node@12.19, npm@6.14.8, yarn@1.22.10
+* Upgrade web-ext@5.3.0, xo@0.34.2, stylelint@13.8.0, ava@3.13.0
+
 # 2.5.1
 * Upgrade web-ext@4.3.0 to fix `web-ext build` env error [#793](https://github.com/mozilla/web-ext/issues/793)